bugfix salvataggi

+++ ore-rd/src/app/progetto-dettaglio/progetto-dettaglio.component.ts
@@ -1,30 +1,19 @@
-import { filter } from 'rxjs/operators';
 import { TipologiaSpesaService } from './../_services/tipospesa.service';
-import { TipologiaSpesaComponent } from './../tipologia-spesa/tipologia-spesa.component';
 import { ProgettiWpService } from './../_services/progetti.wp.service';
 import { ProgettiSpesaService } from './../_services/progetti.spesa.service';
 import { MatTableDataSource } from '@angular/material/table';
-import { TipoCosto } from './../_models/tipocosto';
-import { Matricola } from './../_models/matricola';
 import { Subscription } from 'rxjs';
-import { Progetto, ProgettoSpesa, ProgettoWp } from './../_models/progetto';
-import { User } from './../_models/user';
+import { Progetto, ProgettoSpesa, ProgettoWp } from './../_models';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ProgettiService } from './../_services/progetti.service';
 import { AlertService } from './../_services/alert.service';
 import { AuthenticationService } from './../_services/authentication.service';
-import { FormGroup, FormControl } from '@angular/forms';
 import { Component, OnInit } from '@angular/core';
-
-
-
 import {MomentDateAdapter, MAT_MOMENT_DATE_ADAPTER_OPTIONS} from '@angular/material-moment-adapter';
 import {DateAdapter, MAT_DATE_FORMATS, MAT_DATE_LOCALE} from '@angular/material/core';
-import {MatDatepicker} from '@angular/material/datepicker';
-import {Moment} from 'moment';
 import * as _moment from 'moment';
 import { formatDate } from '@angular/common';
-import { PageEvent } from '@angular/material/paginator';
+
 export const MY_FORMATS = {
   parse: {
       dateInput: 'LL'
@@ -55,15 +44,16 @@ export class ProgettoDettaglioComponent implements OnInit {
   progettoWp!: ProgettoWp[];
   progetto_old!: Progetto;
   displayedColumns: string[] = ['descrizione','importo', 'tipologia', 'actions'];
-  displayedColumnsWp: string[] = ['id','titolo', 'descrizione', 'dataInizio', 'dataFine','risorse','monteOre', 'actions'];
-  dataSource = new MatTableDataSource<[]>();
-  dataSourceWp = new MatTableDataSource<[]>();
+  displayedColumnsWp: string[] = ['id','titolo', 'descrizione', 'dataInizio', 'dataFine', 'risorse', 'monteOre', 'mesiUomo', 'actions'];
+  dataSource = new MatTableDataSource<ProgettoSpesa>();
+  dataSourceWp = new MatTableDataSource<ProgettoWp>();
   allTipologie: any[] = [];
   allMatricole: any[] = [];
   allTipiCosto: any[] = [];
   isNotAnnulable:boolean = false;
   id_progetto!: any;
-  errore_stringa= '';
+  errore_stringa = '';
+  MONTE_ORE_MENSILE_PREVISTO = 1720 / 12; // 143.3333
 
   constructor(private authenticationService: AuthenticationService,
     private progettiService: ProgettiService,
@@ -75,9 +65,9 @@ export class ProgettoDettaglioComponent implements OnInit {
     private router: Router) {
 
       this.projectSubscription = this.route.params.subscribe(params => {
-        if(params['id_progetto'] == 'nuovo'){
+        if (params['id_progetto'] == 'nuovo') {
           this.id_progetto = null; 
-        }else{
+        } else {
           this.id_progetto = +params['id_progetto']; 
         }
       },
@@ -88,7 +78,7 @@ export class ProgettoDettaglioComponent implements OnInit {
 
 
   ngOnInit(): void {
-    if(this.id_progetto != null){
+    if (this.id_progetto != null) {
       this.getProgetto();
       this.getProgettoSpesa();
       this.getTipoSpesa();
@@ -103,129 +93,138 @@ export class ProgettoDettaglioComponent implements OnInit {
   getProgettowP(): void {
     this.progettiWpService.getById(this.id_progetto)
       .subscribe(response => {
-        this.progettoWp = response["value"];
-        this.dataSourceWp = new MatTableDataSource<[]>(response["value"]);
+        this.progettoWp = response.data;
+        this.dataSourceWp = new MatTableDataSource(response.data);
       },
       error => {
-        this.dataSourceWp = new MatTableDataSource<[]>();
+        this.dataSourceWp = new MatTableDataSource();
       });
   }
 
   getProgettoSpesa(): void {
     this.progettiSpesaService.getById(this.id_progetto)
       .subscribe(response => {
-        this.dataSource = new MatTableDataSource<[]>(response["value"]);
+        this.dataSource = new MatTableDataSource(response.data);
       },
       error => {
-        this.dataSource = new MatTableDataSource<[]>();
+        this.dataSource = new MatTableDataSource();
       });
   }
+
   getProgetto(): void {
     this.progettiService.getById(this.id_progetto)
       .subscribe(response => {
         this.progetto = new Progetto;
-        this.progetto = response["value"];
-        this.progetto_old = response["value"];
+        this.progetto = response.value;
+        this.progetto_old = response.value;
       },
       error => {
         this.alertService.error(error);
       });
   }
 
-  getRecord(prgSpesa: ProgettoSpesa){
+  getRecord(prgSpesa: ProgettoSpesa) {
+    if (this.dataSource.data) {
+      this.dataSource.data.forEach(x => x.isEditable = false);
+    }
     prgSpesa.isEditable = true;
   }
 
-  getRecordwP(prgWp: ProgettoWp){
+  getRecordwP(prgWp: ProgettoWp) {
+    if (this.dataSourceWp.data) {
+      this.dataSourceWp.data.forEach(x => x.isEditable = false);
+    }
     prgWp.isEditable = true;
   }
 
   getMatricole(): void {
     this.progettiService.getAllMatricole()
       .subscribe(response => {
-        this.allMatricole = response["data"];
-      },
-      error => {
-        this.alertService.error(error);
-      });
-  }
- stampaRisorse(risorse: string[]){
-    let arrayRisorse:any[] = [];
-    risorse.forEach(element => {
-      let  u  = this.allMatricole.find(x => x.MATRICOLA == element);
-      if(this.allMatricole.find(x => x.MATRICOLA == element) != null)
-        arrayRisorse.push(this.allMatricole.find(x => x.MATRICOLA == element).NOME);
-    });
-    return arrayRisorse;
- }
-  getSupervisor(): void {
-    this.progettiService.getAllTipiCostoPanthera()
-      .subscribe(response => {
-        this.allTipiCosto = response["data"];
+        this.allMatricole = response.data;
       },
       error => {
         this.alertService.error(error);
       });
   }
 
-  getTipoSpesa(){
+ stampaRisorse(risorse: string[]) {
+    let arrayRisorse:any[] = [];
+    risorse.forEach(element => {
+      let  u  = this.allMatricole.find(x => x.MATRICOLA == element);
+      if (this.allMatricole.find(x => x.MATRICOLA == element) != null)
+        arrayRisorse.push(this.allMatricole.find(x => x.MATRICOLA == element).NOME);
+    });
+    return arrayRisorse;
+ }
+
+  getSupervisor(): void {
+    this.progettiService.getAllTipiCostoPanthera()
+      .subscribe(response => {
+        this.allTipiCosto = response.data;
+      },
+      error => {
+        this.alertService.error(error);
+      });
+  }
+
+  getTipoSpesa() {
     this.tipologiaSpesaService.getAll()
         .subscribe(response => {
-          this.allTipologie = response["data"];
+          this.allTipologie = response.data;
         },
         error => {
         });
   }
 
   salva() {
-    
-    if(this.progetto.DATA_FINE)
+    if (this.progetto.DATA_FINE)
       this.progetto.DATA_FINE = formatDate(this.progetto.DATA_FINE,"YYYY-MM-dd","en-GB");
 
-    if(this.progetto.DATA_INIZIO)
+    if (this.progetto.DATA_INIZIO)
       this.progetto.DATA_INIZIO = formatDate(this.progetto.DATA_INIZIO,"YYYY-MM-dd","en-GB");
     
-    if(this.id_progetto == null){
+    if (this.id_progetto == null) {
       this.progettiService.insert(this.progetto)
       .subscribe(response => {
         this.alertService.success("Progetto inserito con successo");
-        this.router.navigate(['/progetto/'+response["value"][0]["ID_PROGETTO"]]);
+        this.router.navigate(['/progetto/' + response.value.ID_PROGETTO]);
       },
       error => {
         this.alertService.error(error);
       });
     } else {
       let no_error = true;
-      let monte_totale_wp:number = 0;
-      for(let i = 0; i < this.progettoWp.length; i++){
-        no_error = this.controlliDate(this.progettoWp[i]);
-        monte_totale_wp = (Number(monte_totale_wp) + Number(this.progettoWp[i].MONTE_ORE));
+      let monte_totale_wp: number = 0;
+      if (this.progettoWp) {
+        for(let i = 0; i < this.progettoWp.length; i++) {
+          no_error = this.controlliDate(this.progettoWp[i]);
+          monte_totale_wp = (Number(monte_totale_wp) + Number(this.progettoWp[i].MONTE_ORE));
+        }
       }
-      if(this.progetto.MONTE_ORE_TOT < monte_totale_wp){
-        this.alertService.error("Il MonteOre dei WP supera quello del Progetto");
+      if (this.progetto.MONTE_ORE_TOT < monte_totale_wp) {
+        this.alertService.error("Il monte ore dei WP supera quello del progetto");
         no_error = false;
       }
-      if(no_error){
+      if (no_error) {
         this.progettiService.update(this.progetto)
         .subscribe(response => {
           this.alertService.success("Progetto modificato con successo");
-          this.router.navigate(['/progetto/'+response["value"]["ID_PROGETTO"]]);
+          this.router.navigate(['/progetto/' + response.value.ID_PROGETTO]);
         },
         error => {
           this.alertService.error(error);
         });
       }
     }
-    
   }
 
   nuovoProgettoSpesa() {  
-    let progettoSpesa_nuovo:any;
+    let progettoSpesa_nuovo: any;
     progettoSpesa_nuovo = {ID_PROGETTO:this.progetto.ID_PROGETTO,ID_SPESA:null, DESCRIZIONE:null,IMPORTO:null,TIPOLOGIA: {ID_TIPOLOGIA:null, DESCRIZIONE:null},isEditable:true,isInsert:true};
     let data:any[] = [];
-    if(this.dataSource.data == null){
+    if (this.dataSource.data == null) {
       data.push(progettoSpesa_nuovo);
-    }else{
+    } else {
       data = this.dataSource.data;
       data.push(progettoSpesa_nuovo);
     }
@@ -233,58 +232,62 @@ export class ProgettoDettaglioComponent implements OnInit {
   } 
 
   nuovoProgettoWp() {  
-    let progettoWp_nuovo:any;
+    let progettoWp_nuovo: ProgettoWp;
     progettoWp_nuovo = {
-      ID_PROGETTO:this.progetto.ID_PROGETTO, 
-      ID_WP:null, 
-      TITOLO:null, 
-      DESCRIZIONE:null, 
+      ID_PROGETTO: this.progetto.ID_PROGETTO, 
+      ID_WP: null, 
+      TITOLO: null, 
+      DESCRIZIONE: null, 
       DATA_INIZIO: null, 
       DATA_FINE: null , 
-      RISORSE:null, 
-      MONTE_ORE: null, 
-      isEditable:true,
-      isInsert:true
+      RISORSE: [], 
+      MONTE_ORE: 0, 
+      isEditable: true,
+      isInsert: true
     };
-    let dataWp:any[] = [];
-    if(this.dataSourceWp.data == null){
+    let dataWp: any[] = [];
+    if (this.dataSourceWp.data == null) {
       this
       dataWp.push(progettoWp_nuovo);
-    }else{
+    } else {
       dataWp = this.dataSourceWp.data;
       dataWp.push(progettoWp_nuovo);
     }
     this.dataSourceWp.data = dataWp;
   } 
 
-  deleteChange(prgSpesa:any){
-    this.progettiSpesaService.delete(prgSpesa.ID_SPESA)
-        .subscribe(response => {
-          this.getProgettoSpesa();
-        },
-        error => {
-          this.alertService.error("Impossibile eliminare il record");
-        });
+  deleteChange(prgSpesa: ProgettoSpesa) {
+    if (prgSpesa.ID_PROGETTO != null && prgSpesa.ID_SPESA != null) {
+      this.progettiSpesaService.delete(prgSpesa.ID_PROGETTO, prgSpesa.ID_SPESA)
+          .subscribe(response => {
+            this.getProgettoSpesa();
+          },
+          error => {
+            this.alertService.error("Impossibile eliminare il record");
+          });
+    }
   }
 
-  deleteChangeWp(prgWp:any){
-    this.progettiWpService.delete(prgWp.ID_WP,prgWp.ID_PROGETTO)
-        .subscribe(response => {
-          this.getProgettowP();
-        },
-        error => {
-          this.alertService.error("La tipologia è stata già utilizzata per un ProgettoSpesa");
-        });
+  deleteChangeWp(prgWp: ProgettoWp) {
+    if (prgWp.ID_WP != null && prgWp.ID_PROGETTO != null) {
+      this.progettiWpService.delete(prgWp.ID_WP, prgWp.ID_PROGETTO)
+      .subscribe(response => {
+        this.getProgettowP();
+      },
+      error => {
+        this.alertService.error("La tipologia è stata già utilizzata per un ProgettoSpesa");
+      });
+    }
   }
   
 
-  salvaModifica(prgSpesa: ProgettoSpesa){
-    if(prgSpesa.ID_SPESA == null){
-      if(prgSpesa.IMPORTO != '' || prgSpesa.IMPORTO != null){
+  salvaModifica(prgSpesa: ProgettoSpesa) {
+    if (prgSpesa.ID_SPESA == null) {
+      if (prgSpesa.IMPORTO != null) {
         this.progettiSpesaService.insert(prgSpesa)
         .subscribe(response => {
           this.dataSource.data.splice(-1, 1);
-          this.dataSource.data.push(response["value"][0]);
+          this.dataSource.data.push(response.value);
           this.dataSource.data = this.dataSource.data;
           this.alertService.success("Spesa salvata con successo");
           prgSpesa.isEditable=false;
@@ -294,12 +297,13 @@ export class ProgettoDettaglioComponent implements OnInit {
         });
       }
     } else {
-      if(prgSpesa.IMPORTO != '' || prgSpesa.IMPORTO != null){
+      if (prgSpesa.IMPORTO != null) {
         this.progettiSpesaService.update(prgSpesa)
         .subscribe(response => {
           this.alertService.success("Spesa modificata con successo");
           this.getProgettoSpesa();
-          prgSpesa.isEditable=false;
+          prgSpesa.isEditable = false;
+          prgSpesa.TIPOLOGIA = response.value.TIPOLOGIA;
         },
         error => {
           this.alertService.error(error);
@@ -308,86 +312,94 @@ export class ProgettoDettaglioComponent implements OnInit {
     }
   }
 
-  controlliDate(wp:any){
+  annullaModifica(row: ProgettoSpesa) {
+    this.getProgettoSpesa();
+  }
+
+  annullaModificaWp(row: ProgettoWp) {
+    this.getProgettowP();
+  }
+
+  controlliDate(wp: ProgettoWp) {
     let datePrIniziale = '';
     let datePrFinale = '';
     let dateWpIniziale = '';
     let dateWpFinale = '';
     this.errore_stringa = '';
 
-    if(this.progetto.DATA_INIZIO){
+    if (this.progetto.DATA_INIZIO) {
       datePrIniziale = formatDate(this.progetto.DATA_INIZIO,'yyyy-MM-dd','en_US');
-    }else{
+    } else {
       this.errore_stringa += "Inserire la Data Inizio Progetto <br/>";
     }
-    if(this.progetto.DATA_FINE){
+    if (this.progetto.DATA_FINE) {
       datePrFinale = formatDate(this.progetto.DATA_FINE,'yyyy-MM-dd','en_US');
-    }else{
+    } else {
         this.errore_stringa += "Inserire la Data Fine Progetto <br/>";
     }
 
-    if(wp.DATA_INIZIO != ''){
+    if (wp.DATA_INIZIO !== null && wp.DATA_INIZIO !== '') {
       dateWpIniziale = formatDate(wp.DATA_INIZIO,'yyyy-MM-dd','en_US');
-    }else{
+    } else {
       this.errore_stringa += "Inserire la Data Inizio Wp <br/>";
     }
 
-    if(wp.DATA_FINE != ''){
+    if (wp.DATA_FINE !== null && wp.DATA_FINE !== '') {
       dateWpFinale = formatDate(wp.DATA_FINE,'yyyy-MM-dd','en_US');
-    }else{
+    } else {
         this.errore_stringa += "Inserire la Data Fine Wp <br/>";
     }
     console.log(datePrIniziale+" > "+dateWpIniziale);
-    if(datePrIniziale != '' && dateWpIniziale != ''){
-      if(datePrIniziale > dateWpIniziale){
+    if (datePrIniziale != '' && dateWpIniziale != '') {
+      if (datePrIniziale > dateWpIniziale) {
         this.errore_stringa = "Un WP non può iniziare prima del Progetto <br/>";
         wp.DATA_INIZIO = '';
       }
     }
     
-    if(datePrFinale != '' && dateWpFinale != ''){
-      if(datePrFinale < dateWpFinale){
+    if (datePrFinale != '' && dateWpFinale != '') {
+      if (datePrFinale < dateWpFinale) {
         this.errore_stringa += "Un WP non può finire dopo il Progetto <br/>";
         wp.DATA_FINE = '';
       }
     }
 
-    if( wp.DATA_INIZIO == '' || wp.DATA_FINE == ''){
+    if ( wp.DATA_INIZIO == '' || wp.DATA_FINE == '') {
       this.alertService.error(this.errore_stringa);
       return false;
     }
     return true;
   }
 
-  salvaModificaWp(prgWp: ProgettoWp){    
+  salvaModificaWp(prgWp: ProgettoWp) {    
     console.log(prgWp);
-    if(prgWp.DATA_FINE)
+    if (prgWp.DATA_FINE)
       prgWp.DATA_FINE = formatDate(prgWp.DATA_FINE,"YYYY-MM-dd","en-GB");
 
-    if(prgWp.DATA_INIZIO)
+    if (prgWp.DATA_INIZIO)
       prgWp.DATA_INIZIO = formatDate(prgWp.DATA_INIZIO,"YYYY-MM-dd","en-GB");
 
     let error = false;
-    if(prgWp.ID_WP == null){
+    if (prgWp.ID_WP == null) {
       let monte_totale_wp = 0;
-      if(this.progettoWp.length > 1){
-        for(let i = 0; i < this.progettoWp.length; i++){
+      if (this.progettoWp.length > 1) {
+        for(let i = 0; i < this.progettoWp.length; i++) {
           monte_totale_wp = (Number(monte_totale_wp) + Number(this.progettoWp[i].MONTE_ORE));
         }
-        if(this.progetto.MONTE_ORE_TOT < monte_totale_wp){
-          this.alertService.error("Il MonteOre dei WP supera quello del Progetto");
+        if (this.progetto.MONTE_ORE_TOT < monte_totale_wp) {
+          this.alertService.error("Il monte ore dei WP supera quello del progetto");
           error = true;
         }
       }
 
-      if(this.controlliDate(prgWp) && !error){
+      if (this.controlliDate(prgWp) && !error) {
         this.progettiWpService.insert(prgWp)
         .subscribe(response => {
           this.alertService.success("Work Package inserito con successo");
           this.dataSourceWp.data.splice(-1, 1);
-          this.dataSourceWp.data.push(response["value"][0]);
+          this.dataSourceWp.data.push(response.value);
           this.dataSourceWp.data = this.dataSourceWp.data;
-          prgWp.isEditable=false;
+          prgWp.isEditable = false;
         },
         error => {
           this.alertService.error(error);
@@ -395,16 +407,16 @@ export class ProgettoDettaglioComponent implements OnInit {
       }
     } else {
       let monte_totale_wp = 0;
-      if(this.progettoWp.length > 1){
-        for(let i = 0; i < this.progettoWp.length; i++){
+      if (this.progettoWp.length > 1) {
+        for(let i = 0; i < this.progettoWp.length; i++) {
           monte_totale_wp = (Number(monte_totale_wp) + Number(this.progettoWp[i].MONTE_ORE));
         }
-        if(this.progetto.MONTE_ORE_TOT < monte_totale_wp){
+        if (this.progetto.MONTE_ORE_TOT < monte_totale_wp) {
           this.alertService.error("Il MonteOre dei WP supera quello del Progetto");
           error = true;
         }
       }
-      if(this.controlliDate(prgWp) && !error){
+      if (this.controlliDate(prgWp) && !error) {
         this.progettiWpService.update(prgWp)
         .subscribe(response => {
           this.alertService.success("Work Package modificato con successo");
@@ -418,9 +430,9 @@ export class ProgettoDettaglioComponent implements OnInit {
   }
   
 
-  undoChange(prgSpesa:ProgettoSpesa){
+  undoChange(prgSpesa: ProgettoSpesa) {
     prgSpesa.isEditable=false;
-    if(prgSpesa.ID_PROGETTO == null){
+    if (prgSpesa.ID_PROGETTO == null) {
       this.dataSource.data.splice(-1, 1);
       this.dataSource.data = this.dataSource.data;
     }
@@ -428,5 +440,19 @@ export class ProgettoDettaglioComponent implements OnInit {
 
   report() {
     this.router.navigate(["progetto", this.progetto.ID_PROGETTO, "report"]);
+  }
+
+  monteOreMensile(monteOre: number|null) {
+    if (monteOre === null) {
+      return null;
+    }
+    return Math.round(monteOre / this.MONTE_ORE_MENSILE_PREVISTO);
+  }
+
+  costoMensile(costoOrario: number|null) {
+    if (costoOrario === null) {
+      return null;
+    }
+    return Math.round(costoOrario * this.MONTE_ORE_MENSILE_PREVISTO);
   }
 }
